chore: remove app settings from schema
